requirements.txt
django
djangorestframework
django-guardian